+++ appWindFlow/1_🌬_Origins.py
@@ -1,0 +1,55 @@
+import streamlit as st
+import base64
+
+import numpy as np
+import pandas as pd
+from PIL import Image
+
+
+
+def add_logo():
+    st.markdown(
+        """
+        <style>
+            [data-testid="stSidebarNav"] {
+                background-image: url(https://i.ibb.co/sHy0KCY/Foto-Jet-9.png);
+                background-repeat: no-repeat;
+                padding-top: 150px;
+                background-position: 0px 15px;
+            }
+            [data-testid="stSidebarNav"]::before {
+                content: "WindFlow 1.0.2";
+                margin-left: 20px;
+                margin-top: 20px;
+                font-size: 15px;
+                position: relative;
+                top: 100px;
+            }
+        </style>
+        """,
+        unsafe_allow_html=True,
+    )
+#####
+add_logo()
+
+
+st.markdown("""# OUR JOURNEY TO WINDFLOW
+## Us 🤓
+We are students from Le wagon For two months we were intensely learning about data science and we culminated these months by working on a joint project, where we put our knowledge to the test and were able to contribute from the areas most explored and known by each of us. """)
+
+image = Image.open('appWindFlow/us.png')
+st.image(image, caption='Sebastián Díaz - John Collins - Sebastián Ossul - France Poma - Felipe Ulloa')
+
+st.sidebar.markdown("# Main Page 🍃")
+st.markdown("""## Problematic: 🚨
+Air pollution problems linked to work in mining stations in the country of Chile.
+Currently there are several localities in the country called "sacrifice zones" where through industries, the lives of the inhabitants of said area are exposed to constant pollution that affects the health and quality of life of people.
+""")
+image = Image.open('appWindFlow/hola.jpg')
+st.image(image, caption='Quintero as a sacrifice zone')
+
+
+st.markdown("""## Solution:💡
+WINDFLOW 🎉
+## What is windflow?: 🌬
+It is an application that helps mining companies to schedule the best time to operate dynamite blasting activities by being able to predict the speed and direction that the wind will have on a given day and time. This reduces the environmental impact generated by the propagation of particulate matter (dust).""")
